Refactor ConfigUpdate\Cache with service test
