--- src/Process/Deploy/InstallUpdate/ConfigUpdate/Cache.php
@@ -1,75 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-namespace Magento\MagentoCloud\Process\Deploy\InstallUpdate\ConfigUpdate;
-
-use Magento\MagentoCloud\Config\Deploy\Reader as ConfigReader;
-use Magento\MagentoCloud\Config\Deploy\Writer as ConfigWriter;
-use Magento\MagentoCloud\Process\Deploy\InstallUpdate\ConfigUpdate\Cache\Config;
-use Magento\MagentoCloud\Process\ProcessInterface;
-use Psr\Log\LoggerInterface;
-
-/**
- * Processes cache configuration.
- */
-class Cache implements ProcessInterface
-{
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
-
-    /**
-     * @var ConfigWriter
-     */
-    private $configWriter;
-
-    /**
-     * @var ConfigReader
-     */
-    private $configReader;
-
-    /**
-     * @var Config
-     */
-    private $cacheConfig;
-
-    /**
-     * @param ConfigReader $configReader
-     * @param ConfigWriter $configWriter
-     * @param LoggerInterface $logger
-     * @param Config $cacheConfig
-     */
-    public function __construct(
-        ConfigReader $configReader,
-        ConfigWriter $configWriter,
-        LoggerInterface $logger,
-        Config $cacheConfig
-    ) {
-        $this->configReader = $configReader;
-        $this->configWriter = $configWriter;
-        $this->logger = $logger;
-        $this->cacheConfig = $cacheConfig;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function execute()
-    {
-        $config = $this->configReader->read();
-        $cacheConfig = $this->cacheConfig->get();
-
-        if (!empty($cacheConfig)) {
-            $this->logger->info('Updating cache configuration.');
-            $config['cache'] = $cacheConfig;
-        } else {
-            $this->logger->info('Cache configuration was not found. Removing cache configuration.');
-            unset($config['cache']);
-        }
-
-        $this->configWriter->create($config);
-    }
-}
